reorganized the isa module in the HAL
