--- kernel/src/hal/isa/implementations/mod.rs
@@ -1,4 +1,0 @@
-#[cfg(target_arch = "aarch64")]
-pub mod aarch64;
-#[cfg(target_arch = "x86_64")]
-pub mod x86_64;
