--- kernel/src/hal/isa/implementations/x86_64/mod.rs
@@ -1,2 +1,0 @@
-pub mod instructions;
-pub mod memory;
