--- kernel/src/hal/isa/interfaces/mod.rs
@@ -1,1 +1,0 @@
-pub mod memory;
