+++ kernel/src/hal/isa/memory/mod.rs
@@ -1,0 +1,5 @@
+pub mod interface;
+#[cfg(target_arch = "x86_64")]
+pub mod x86_64;
+#[cfg(target_arch = "aarch64")]
+pub mod aarch64;
