+++ kernel/src/hal/isa/memory/x86_64/instructions.rs
@@ -3,7 +3,7 @@ use core::arch::asm;
 pub unsafe fn outb(port: u16, value: u8) {
     unsafe {
         asm!(
-            "out dx, al", 
+            "out dx, al",
             in("dx") port, in("al") value
         );
     }
@@ -13,7 +13,7 @@ pub unsafe fn inb(port: u16) -> u8 {
     let value: u8;
     unsafe {
         asm!(
-            "in al, dx", 
+            "in al, dx",
             out("al") value, in("dx") port
         );
     }
