+++ kernel/src/hal/isa/memory/x86_64/memory/address/mod.rs
@@ -1,0 +1,15 @@
+pub mod ioaddr;
+pub mod paddr;
+pub mod vaddr;
+
+use super::instructions::{inb, outb};
+use crate::hal::isa::interfaces::memory;
+
+pub struct Memory;
+
+impl crate::hal::isa::interfaces::memory::Memory for Memory {
+    type VAddr = vaddr::VAddr;
+    type PAddr = PAddr;
+    type IoAddr = IoAddr;
+}
+
