+++ kernel/src/hal/isa/memory/x86_64/memory/address/paddr.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+struct PAddr {
+    addr: u64
+}
